Move ClassificationTable component to classification folder
